Awake instead of start means we don't need to wait until second frame
gonna keep it in GMs coroutine, because we should add the card, let a frame pass, then call shuffle maybe.
That or make shuffle in the deck a coroutine and pass a frame every like 20 cards or something. Realistically I can't imagine less than 200 cards (which it will almost certainly be) being an issue, but best practice may be to do it over a few frames or in a thread.